pyproject.toml
[project]
name = "qkb-trial"
version = "0.1.0"
description = "CLI tool for managing and querying internal knowledge files via SQLite FTS5 + optional LLM rerank."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.7",
  "httpx>=0.27",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.5"]

[project.scripts]
qkb = "qkb.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qkb"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
